Moved files to make detectors folder
